backend/kyc-service/requirements.txt
# PayU KYC Service - Python FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
alembic==1.14.0

# Kafka
aiokafka==0.11.0

# OCR
paddlepaddle==3.0.0
paddleocr==2.9.0
paddle2onnx==1.3.0
# Alternative: tesseract-ocr

# Image Processing
opencv-python==4.10.0
numpy==1.26.4
pillow==10.4.0

# Face Recognition & Liveness
face-recognition==1.3.0
dlib==19.24.6
scikit-learn==1.5.1
torch==2.5.0
torchvision==0.20.0

# HTTP Client
httpx==0.27.0
aiohttp==3.10.0

# Utilities
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Logging & Monitoring
structlog==24.4.0
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-httpx==0.48b0
opentelemetry-exporter-otlp==1.27.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.0
